pyproject.toml
[tool.poetry]
name = "helpers"
version = "0.1.0"
description = ""
authors = ["Михаил <mxoilslukin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12"
fastapi = "^0.115.8"
pydantic = "^2.10.6"
aiokafka = "^0.12.0"
sqlalchemy = "^2.0.37"
factory-boy = "^3.3.3"
jose = "^1.0.0"
pytest = "^8.3.4"
orjson = "^3.10.15"
httpx = "^0.28.1"
loguru = "^0.7.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
